% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holding_function.R
\name{holding_function}
\alias{holding_function}
\title{A slightly modified version of the knockoff function}
\usage{
holding_function(resp, expl, Xko, offset = 0, fdr = 0.2, cores = 2, ...)
}
\arguments{
\item{resp}{The response variable, a matrix the same length as expl The first col must be the identifier}

\item{expl}{The data frame to be the explanatory variables, the first col must be the identifier}

\item{Xko}{The knockoff variables created seperately (see the 'knockoff' package for more details). Also must be the same length as expl, resp.}

\item{...}{Various other arguments to be passed to the my_kn function.}
}
\value{
This will return of data frame 5 columns. The first being phenotype, the signficant variable selected, the estimate of a glm regression, se for standard error of the regression and the p-value.
}
\description{
2 major differences being that the knockoff variables
are to be passed directly to the function, not made inside
of it; and secondly you can pass the amount of cores wanted
directly to the statistic function chosen. Be careful as this
may need to be 'NULL'-ed if the statistic function you choose
does not allow for multi-cores.
}
