% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_function.R
\name{pbd_kn}
\alias{pbd_kn}
\title{Access to pbdMPI in use of matrix_kn among mulitple files.}
\usage{
pbd_kn(dir, ...)
}
\arguments{
\item{dir}{This is the directory where the folders (AND ONLY THE FOLDERS) are located}

\item{...}{Any other argument to be passed to the matrix_kn function, and in turn my_kn}
}
\description{
Major thing is that we are assuming we have a directory with any number of folders. 
Inside each of these folders there needs to be two files,
one called "resp" with the response variables and one called 
"expl" for the explanatory variables. The folders will be divided among
the nodes and each node will do the analysis of it's folders,
recombining them at the end.
}
