% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_function.R
\name{pbd_kn}
\alias{pbd_kn}
\title{Access to pbdMPI in use of matrix_kn among mulitple files.}
\usage{
pbd_kn(dir, ...)
}
\arguments{
\item{dir}{This is the directory where the folders (AND ONLY THE FOLDERS) are located}

\item{...}{Any other argument to be passed to the matrix_kn function, and in turn my_kn}
}
\value{
One large data frame of 6 columns 
\itemize{
\item{Resp: } What ever response variable was being tested
\item{Expl: } Gives the name of the signficant explanatory variable chosen
\item{Est: } Estimate of the glm model, defaults to normal
\item{se: } Gives the standard error of the parameter estimate from the glm
\item{p-value: } Gives the p-value associated with that variable estimate from the glm for the null hypothesis the effect magnitude is 0.
\item{file: } Which file the data was used from
}
}
\description{
Major thing is that we are assuming we have a directory with any number of folders. 
Inside each of these folders there needs to be two files,
one called "resp" with the response variables and one called 
"expl" for the explanatory variables. The folders will be divided among
the nodes and each node will do the analysis of it's folders,
recombining them at the end.
}
